requirements.txt
Flask
flask_wtf
wtforms
flask_sqlalchemy
flask_migrate
requests